Make the resolveFile struct "unix" instead of out of "darwin"-only.
No need to maintain two parses for this kind of file.
