+++ pkg/client/rootd/dns/server_darwin.go
@@ -1,14 +1,11 @@
 package dns
 
 import (
-	"bufio"
 	"bytes"
 	"context"
-	"fmt"
 	"net"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 
 	"github.com/datawire/dlib/dgroup"
@@ -16,97 +13,6 @@ import (
 	"github.com/telepresenceio/telepresence/v2/pkg/maps"
 	"github.com/telepresenceio/telepresence/v2/pkg/vif"
 )
-
-type resolveFile struct {
-	port        int
-	domain      string
-	nameservers []net.IP
-	search      []string
-}
-
-func readResolveFile(fileName string) (*resolveFile, error) {
-	fl, err := os.Open(fileName)
-	if err != nil {
-		return nil, err
-	}
-	defer fl.Close()
-	sc := bufio.NewScanner(fl)
-	rf := resolveFile{}
-	line := 0
-
-	onlyOne := func(key string) error {
-		return fmt.Errorf("%q must have a value at %s line %d", key, fileName, line)
-	}
-
-	for sc.Scan() {
-		line++
-		txt := strings.TrimSpace(sc.Text())
-		if len(txt) == 0 || strings.HasPrefix(txt, "#") {
-			continue
-		}
-		fields := strings.Fields(txt)
-		fc := len(fields)
-		if fc == 0 {
-			continue
-		}
-		key := fields[0]
-		if fc == 1 {
-			return nil, fmt.Errorf("%q must have a value at %s line %d", key, fileName, line)
-		}
-		value := fields[1]
-		switch key {
-		case "port":
-			if fc != 2 {
-				return nil, onlyOne(key)
-			}
-			rf.port, err = strconv.Atoi(value)
-			if err != nil {
-				return nil, fmt.Errorf("%q is not a valid integer at %s line %d", key, fileName, line)
-			}
-		case "domain":
-			if fc != 2 {
-				return nil, onlyOne(key)
-			}
-			rf.domain = value
-		case "nameserver":
-			if fc != 2 {
-				return nil, onlyOne(key)
-			}
-			ip := net.ParseIP(value)
-			if ip == nil {
-				return nil, fmt.Errorf("value %q for %q is not a valid IP at %s line %d", value, key, fileName, line)
-			}
-			rf.nameservers = append(rf.nameservers, ip)
-		case "search":
-			rf.search = fields[1:]
-		default:
-			// This reader doesn't do options just yet
-			return nil, fmt.Errorf("%q is not a recognized key at %s line %d", key, fileName, line)
-		}
-	}
-	return &rf, nil
-}
-
-func (r *resolveFile) write(fileName string) error {
-	buf := bytes.NewBufferString("# Generated by telepresence\n")
-	fmt.Fprintf(buf, "port %d\n", r.port)
-	if r.domain != "" {
-		fmt.Fprintf(buf, "domain %s\n", r.domain)
-	}
-	for _, ns := range r.nameservers {
-		fmt.Fprintf(buf, "nameserver %s\n", ns)
-	}
-
-	if len(r.search) > 0 {
-		buf.WriteString("search")
-		for _, s := range r.search {
-			buf.WriteByte(' ')
-			buf.WriteString(s)
-		}
-		buf.WriteByte('\n')
-	}
-	return os.WriteFile(fileName, buf.Bytes(), 0o644)
-}
 
 func (r *resolveFile) setSearchPaths(paths ...string) {
 	ps := make([]string, 0, len(paths)+1)
@@ -118,6 +24,12 @@ func (r *resolveFile) setSearchPaths(paths ...string) {
 	}
 	ps = append(ps, r.domain)
 	r.search = ps
+}
+
+func (r *resolveFile) write(fileName string) error {
+	var buf bytes.Buffer
+	_, _ = r.WriteTo(&buf)
+	return os.WriteFile(fileName, buf.Bytes(), 0o644)
 }
 
 // Worker places a file under the /etc/resolver directory so that it is picked up by the
@@ -162,7 +74,7 @@ func (s *Server) Worker(c context.Context, dev vif.Device, configureDNS func(net
 	rf := resolveFile{
 		port:        dnsAddr.Port,
 		domain:      kubernetesZone,
-		nameservers: []net.IP{dnsAddr.IP},
+		nameservers: []string{dnsAddr.IP.String()},
 		search:      []string{kubernetesZone},
 	}
 	if err = rf.write(resolverFileName); err != nil {
@@ -271,7 +183,7 @@ func (s *Server) updateResolverFiles(c context.Context, resolverDirName, resolve
 		df := resolveFile{
 			port:        dnsAddr.Port,
 			domain:      domain,
-			nameservers: []net.IP{dnsAddr.IP},
+			nameservers: []string{dnsAddr.IP.String()},
 		}
 		nsFile := domainResolverFile(resolverDirName, domain)
 		dlog.Infof(c, "Generated new %s", nsFile)
